Move core locale handling to avoid impact settings package size
